+++ src/document/vector_store.mock.js
@@ -15,31 +15,31 @@ export class MockVectorStore {
     }
 
     add(document) {
-        return this.#ctrl.call(this, "add", document);
+        return this.#ctrl.call(this, 'add', document);
     }
 
     addBatch(documents) {
-        return this.#ctrl.call(this, "addBatch", documents);
+        return this.#ctrl.call(this, 'addBatch', documents);
     }
 
     clear() {
-        return this.#ctrl.call(this, "clear");
+        return this.#ctrl.call(this, 'clear');
     }
 
     delete(id) {
-        return this.#ctrl.call(this, "delete", id);
+        return this.#ctrl.call(this, 'delete', id);
     }
 
     findSimilar(queryVectors) {
-        return this.#ctrl.call(this, "findSimilar", queryVectors);
+        return this.#ctrl.call(this, 'findSimilar', queryVectors);
     }
 
     size() {
-        return this.#ctrl.call(this, "size");
+        return this.#ctrl.call(this, 'size');
     }
 
     stats() {
-        return this.#ctrl.call(this, "stats");
+        return this.#ctrl.call(this, 'stats');
     }
 }
 
@@ -49,37 +49,37 @@ export class MockVectorStoreRecorder {
 
     constructor(ctrl, mock) {
         if (!(mock instanceof MockVectorStore)) {
-            throw new Error("mock must be an instance of MockVectorStore");
+            throw new Error('mock must be an instance of MockVectorStore');
         }
         this.#ctrl = ctrl;
         this.#mock = mock;
     }
 
     add(document) {
-        return this.#ctrl.recordCall(this.#mock, "add", document);
+        return this.#ctrl.recordCall(this.#mock, 'add', document);
     }
 
     addBatch(documents) {
-        return this.#ctrl.recordCall(this.#mock, "addBatch", documents);
+        return this.#ctrl.recordCall(this.#mock, 'addBatch', documents);
     }
 
     clear() {
-        return this.#ctrl.recordCall(this.#mock, "clear");
+        return this.#ctrl.recordCall(this.#mock, 'clear');
     }
 
     delete(id) {
-        return this.#ctrl.recordCall(this.#mock, "delete", id);
+        return this.#ctrl.recordCall(this.#mock, 'delete', id);
     }
 
     findSimilar(queryVectors) {
-        return this.#ctrl.recordCall(this.#mock, "findSimilar", queryVectors);
+        return this.#ctrl.recordCall(this.#mock, 'findSimilar', queryVectors);
     }
 
     size() {
-        return this.#ctrl.recordCall(this.#mock, "size");
+        return this.#ctrl.recordCall(this.#mock, 'size');
     }
 
     stats() {
-        return this.#ctrl.recordCall(this.#mock, "stats");
+        return this.#ctrl.recordCall(this.#mock, 'stats');
     }
 }